# Add support for compact limbs (2 limb mode) to multi-range-check gadget