--- a/kimchi/src/circuits/polynomials/range_check/circuitgates.rs
+++ b/kimchi/src/circuits/polynomials/range_check/circuitgates.rs
@@ -58,25 +58,25 @@
 //~ * `vipj` is the jth 12-bit limb of value $v_i$
 //~ * `vicj` is the jth 2-bit crumb limb of value $v_i$
 //~
-//~ | Gates | `RangeCheck0`  | `RangeCheck0`  | `RangeCheck1`  | `Zero`          |
-//~ | ----- | -------------- | -------------- | -------------- | --------------- |
-//~ | Rows  |          0     |          1     |          2     |          3      |
-//~ | Cols  |                |                |                |                 |
-//~ |     0 |         `v0`   |         `v1`   |         `v2`   |         `0`     |
-//~ |  MS:1 | copy    `v0p0` | copy    `v1p0` | crumb   `v2c0` | crumb   `v2c10` |
-//~ |     2 | copy    `v0p1` | copy    `v1p1` | crumb   `v2c1` | crumb   `v2c11` |
-//~ |     3 | plookup `v0p2` | plookup `v1p2` | plookup `v2p0` | plookup `v0p0`  |
-//~ |     4 | plookup `v0p3` | plookup `v1p3` | plookup `v2p1` | plookup `v0p1`  |
-//~ |     5 | plookup `v0p4` | plookup `v1p4` | plookup `v2p2` | plookup `v1p0`  |
-//~ |     6 | plookup `v0p5` | plookup `v1p5` | plookup `v2p3` | plookup `v1p1`  |
-//~ |     7 | crumb   `v0c0` | crumb   `v1c0` | crumb   `v2c2` | crumb   `v2c12` |
-//~ |     8 | crumb   `v0c1` | crumb   `v1c1` | crumb   `v2c3` | crumb   `v2c13` |
-//~ |     9 | crumb   `v0c2` | crumb   `v1c2` | crumb   `v2c4` | crumb   `v2c14` |
-//~ |    10 | crumb   `v0c3` | crumb   `v1c3` | crumb   `v2c5` | crumb   `v2c15` |
-//~ |    11 | crumb   `v0c4` | crumb   `v1c4` | crumb   `v2c6` | crumb   `v2c16` |
-//~ |    12 | crumb   `v0c5` | crumb   `v1c5` | crumb   `v2c7` | crumb   `v2c17` |
-//~ |    13 | crumb   `v0c6` | crumb   `v1c6` | crumb   `v2c8` | crumb   `v2c18` |
-//~ | LS:14 | crumb   `v0c7` | crumb   `v1c7` | crumb   `v2c9` | crumb   `v2c19` |
+//~ | Gates | `RangeCheck0`  | `RangeCheck0`  | `RangeCheck1`   | `Zero`          |
+//~ | ----- | -------------- | -------------- | --------------- | --------------- |
+//~ | Rows  |          0     |          1     |          2      |          3      |
+//~ | Cols  |                |                |                 |                 |
+//~ |     0 |         `v0`   |         `v1`   |          `v2`   | crumb   `v2c9`  |
+//~ |  MS:1 | copy    `v0p0` | copy    `v1p0` | optional `v12`  | crumb   `v2c10` |
+//~ |     2 | copy    `v0p1` | copy    `v1p1` | crumb    `v2c0` | crumb   `v2c11` |
+//~ |     3 | plookup `v0p2` | plookup `v1p2` | plookup  `v2p0` | plookup `v0p0`  |
+//~ |     4 | plookup `v0p3` | plookup `v1p3` | plookup  `v2p1` | plookup `v0p1`  |
+//~ |     5 | plookup `v0p4` | plookup `v1p4` | plookup  `v2p2` | plookup `v1p0`  |
+//~ |     6 | plookup `v0p5` | plookup `v1p5` | plookup  `v2p3` | plookup `v1p1`  |
+//~ |     7 | crumb   `v0c0` | crumb   `v1c0` | crumb    `v2c1` | crumb   `v2c12` |
+//~ |     8 | crumb   `v0c1` | crumb   `v1c1` | crumb    `v2c2` | crumb   `v2c13` |
+//~ |     9 | crumb   `v0c2` | crumb   `v1c2` | crumb    `v2c3` | crumb   `v2c14` |
+//~ |    10 | crumb   `v0c3` | crumb   `v1c3` | crumb    `v2c4` | crumb   `v2c15` |
+//~ |    11 | crumb   `v0c4` | crumb   `v1c4` | crumb    `v2c5` | crumb   `v2c16` |
+//~ |    12 | crumb   `v0c5` | crumb   `v1c5` | crumb    `v2c6` | crumb   `v2c17` |
+//~ |    13 | crumb   `v0c6` | crumb   `v1c6` | crumb    `v2c7` | crumb   `v2c18` |
+//~ | LS:14 | crumb   `v0c7` | crumb   `v1c7` | crumb    `v2c8` | crumb   `v2c19` |
 //~
 //~ The 12-bit chunks are constrained with plookups and the 2-bit crumbs are
 //~ constrained with degree-4 constraints of the form $x (x - 1) (x - 2) (x - 3)$.
@@ -155,7 +155,7 @@ where
     F: PrimeField,
 {
     const ARGUMENT_TYPE: ArgumentType = ArgumentType::Gate(GateType::RangeCheck0);
-    const CONSTRAINTS: u32 = 9;
+    const CONSTRAINTS: u32 = 10;
 
     // Constraints for RangeCheck0
     //   * Operates on Curr row
@@ -202,6 +202,15 @@ where
         // Check value v against the sum of limbs
         constraints.push(sum_of_limbs - env.witness_curr(0));
 
+        // Optional compact limbs format (enabled when coeff[0] == 1, disabled when coeff[1] = 0)
+        //   Constrain decomposition of compact limb next(1)
+        //   next(1) = curr(0) + 2^L * next(0)
+        constraints.push(
+            env.coeff(0)
+                * (env.witness_next(1)
+                    - (env.witness_curr(0) + T::two_to_limb() * env.witness_next(0))),
+        );
+
         constraints
     }
 }
@@ -218,23 +227,23 @@ where
 //~
 //~ Given value `v2` the layout looks like this
 //~
-//~ | Column | `Curr`         | `Next`        |
-//~ | ------ | -------------- | ------------- |
-//~ |      0 |         `v2`   | (ignored)     |
-//~ |      1 | crumb   `v2c0` | crumb `v2c10` |
-//~ |      2 | crumb   `v2c1` | crumb `v2c11` |
-//~ |      3 | plookup `v2p0` | (ignored)     |
-//~ |      4 | plookup `v2p1` | (ignored)     |
-//~ |      5 | plookup `v2p2` | (ignored)     |
-//~ |      6 | plookup `v2p3` | (ignored)     |
-//~ |      7 | crumb   `v2c2` | crumb `v2c12` |
-//~ |      8 | crumb   `v2c3` | crumb `v2c13` |
-//~ |      9 | crumb   `v2c4` | crumb `v2c14` |
-//~ |     10 | crumb   `v2c5` | crumb `v2c15` |
-//~ |     11 | crumb   `v2c6` | crumb `v2c16` |
-//~ |     12 | crumb   `v2c7` | crumb `v2c17` |
-//~ |     13 | crumb   `v2c8` | crumb `v2c18` |
-//~ |     14 | crumb   `v2c9` | crumb `v2c19` |
+//~ | Column | `Curr`          | `Next`        |
+//~ | ------ | --------------- | ------------- |
+//~ |      0 |          `v2`   | crumb `v2c9`  |
+//~ |      1 | optional `v12`  | crumb `v2c10` |
+//~ |      2 | crumb    `v2c0` | crumb `v2c11` |
+//~ |      3 | plookup  `v2p0` | (ignored)     |
+//~ |      4 | plookup  `v2p1` | (ignored)     |
+//~ |      5 | plookup  `v2p2` | (ignored)     |
+//~ |      6 | plookup  `v2p3` | (ignored)     |
+//~ |      7 | crumb    `v2c1` | crumb `v2c12` |
+//~ |      8 | crumb    `v2c2` | crumb `v2c13` |
+//~ |      9 | crumb    `v2c3` | crumb `v2c14` |
+//~ |     10 | crumb    `v2c4` | crumb `v2c15` |
+//~ |     11 | crumb    `v2c5` | crumb `v2c16` |
+//~ |     12 | crumb    `v2c6` | crumb `v2c17` |
+//~ |     13 | crumb    `v2c7` | crumb `v2c18` |
+//~ |     14 | crumb    `v2c8` | crumb `v2c19` |
 //~
 //~ where the notation `v2ci` and `v2pi` defined in the "Layout" section above.
 
@@ -254,10 +263,9 @@ where
     //   * Constrain that combining all limbs equals the value v2 stored in row Curr, column 0
     fn constraint_checks<T: ExprOps<F>>(env: &ArgumentEnv<F, T>) -> Vec<T> {
         // 1) Apply range constraints on limbs for Curr row
-        //    * Columns 1-2 are 2-bit crumbs
-        let mut constraints = (1..=2)
-            .map(|i| crumb(&env.witness_curr(i)))
-            .collect::<Vec<T>>();
+        //    * Column 2 is a 2-bit crumb
+        let mut constraints = vec![crumb(&env.witness_curr(2))];
+
         //    * Columns 3-6 are 12-bit plookup range constraints (these are specified
         //      in the lookup gate)
         //    * Columns 7-14 are 2-bit crumb range constraints
@@ -268,9 +276,9 @@ where
         );
 
         // 2) Apply range constraints on limbs for Next row
-        //    * Columns 1-2 are 2-bit crumbs
+        //    * Columns 0-2 are 2-bit crumbs
         constraints.append(
-            &mut (1..=2)
+            &mut (0..=2)
                 .map(|i| crumb(&env.witness_next(i)))
                 .collect::<Vec<T>>(),
         );
@@ -305,7 +313,7 @@ where
         }
 
         // Next row:  Sum remaining 2-bit limbs vc10 and vc11
-        for i in (1..=2).rev() {
+        for i in (0..=2).rev() {
             sum_of_limbs += power_of_2.clone() * env.witness_next(i);
             power_of_2 *= 4u64.into(); // 2 bits
         }
@@ -322,11 +330,8 @@ where
             power_of_2 *= 4096u64.into(); // 12 bits
         }
 
-        // Curr row:  Sum remaining 2-bit limbs: vc0 and vc1
-        for i in (1..=2).rev() {
-            sum_of_limbs += power_of_2.clone() * env.witness_curr(i);
-            power_of_2 *= 4u64.into(); // 2 bits
-        }
+        // Curr row:  Add remaining 2-bit limb v2c0 to sum
+        sum_of_limbs += power_of_2.clone() * env.witness_curr(2);
 
         // Check value v2 against the sum of limbs
         constraints.push(sum_of_limbs - env.witness_curr(0));