--- a/kimchi/src/circuits/polynomials/range_check/gadget.rs
+++ b/kimchi/src/circuits/polynomials/range_check/gadget.rs
@@ -38,9 +38,72 @@ impl<F: PrimeField + SquareRootField> CircuitGate<F> {
     ///       `next_row`      - next row after this gate
     ///       `circuit_gates` - vector of circuit gates comprising this gate
     pub fn create_multi_range_check(start_row: usize) -> (usize, Vec<Self>) {
+        Self::create_multi_range_check_gadget(start_row, false)
+    }
+
+    /// Create range check gate for constraining compact limbs.
+    ///     Inputs the starting row
+    ///     Outputs tuple (`next_row`, `circuit_gates`) where
+    ///       `next_row`      - next row after this gate
+    ///       `circuit_gates` - vector of circuit gates comprising this gate
+    pub fn create_compact_multi_range_check(start_row: usize) -> (usize, Vec<Self>) {
+        Self::create_multi_range_check_gadget(start_row, true)
+    }
+
+    /// Create foreign field muti-range-check gadget by extending the existing gates
+    pub fn extend_multi_range_check(gates: &mut Vec<Self>, curr_row: &mut usize) {
+        let (next_row, circuit_gates) = Self::create_multi_range_check(*curr_row);
+        *curr_row = next_row;
+        gates.extend_from_slice(&circuit_gates);
+    }
+
+    /// Create foreign field muti-range-check gadget by extending the existing gates
+    pub fn extend_compact_multi_range_check(gates: &mut Vec<Self>, curr_row: &mut usize) {
+        let (next_row, circuit_gates) = Self::create_compact_multi_range_check(*curr_row);
+        *curr_row = next_row;
+        gates.extend_from_slice(&circuit_gates);
+    }
+
+    /// Create single range check gate
+    ///     Inputs the starting row
+    ///     Outputs tuple (`next_row`, `circuit_gates`) where
+    ///       `next_row`      - next row after this gate
+    ///       `circuit_gates` - vector of circuit gates comprising this gate
+    pub fn create_range_check(start_row: usize) -> (usize, Vec<Self>) {
+        let gate = CircuitGate::new(
+            GateType::RangeCheck0,
+            Wire::for_row(start_row),
+            vec![F::zero()],
+        );
+        (start_row + 1, vec![gate])
+    }
+
+    /// Create foreign field range-check gate by extending the existing gates
+    pub fn extend_range_check(gates: &mut Vec<Self>, curr_row: &mut usize) {
+        let (next_row, circuit_gates) = Self::create_range_check(*curr_row);
+        *curr_row = next_row;
+        gates.extend_from_slice(&circuit_gates);
+    }
+
+    // Create range check gate for constraining three 88-bit values.
+    //     Inputs the starting row and whether the limbs are in compact format
+    //     Outputs tuple (`next_row`, `circuit_gates`) where
+    //       `next_row`      - next row after this gate
+    //       `circuit_gates` - vector of circuit gates comprising this gate
+    fn create_multi_range_check_gadget(start_row: usize, compact: bool) -> (usize, Vec<Self>) {
+        let coeff = if compact { F::one() } else { F::zero() };
+
         let mut circuit_gates = vec![
-            CircuitGate::new(GateType::RangeCheck0, Wire::for_row(start_row), vec![]),
-            CircuitGate::new(GateType::RangeCheck0, Wire::for_row(start_row + 1), vec![]),
+            CircuitGate::new(
+                GateType::RangeCheck0,
+                Wire::for_row(start_row),
+                vec![F::zero()],
+            ),
+            CircuitGate::new(
+                GateType::RangeCheck0,
+                Wire::for_row(start_row + 1),
+                vec![coeff],
+            ),
             CircuitGate::new(GateType::RangeCheck1, Wire::for_row(start_row + 2), vec![]),
             CircuitGate::new(GateType::Zero, Wire::for_row(start_row + 3), vec![]),
         ];
@@ -58,30 +121,6 @@ impl<F: PrimeField + SquareRootField> CircuitGate<F> {
         circuit_gates.connect_cell_pair((1, 2), (3, 6));
 
         (start_row + circuit_gates.len(), circuit_gates)
-    }
-
-    /// Create foreign field muti-range-check gadget by extending the existing gates
-    pub fn extend_multi_range_check(gates: &mut Vec<Self>, curr_row: &mut usize) {
-        let (next_row, circuit_gates) = Self::create_multi_range_check(*curr_row);
-        *curr_row = next_row;
-        gates.extend_from_slice(&circuit_gates);
-    }
-
-    /// Create single range check gate
-    ///     Inputs the starting row
-    ///     Outputs tuple (`next_row`, `circuit_gates`) where
-    ///       `next_row`      - next row after this gate
-    ///       `circuit_gates` - vector of circuit gates comprising this gate
-    pub fn create_range_check(start_row: usize) -> (usize, Vec<Self>) {
-        let gate = CircuitGate::new(GateType::RangeCheck0, Wire::for_row(start_row), vec![]);
-        (start_row + 1, vec![gate])
-    }
-
-    /// Create foreign field range-check gate by extending the existing gates
-    pub fn extend_range_check(gates: &mut Vec<Self>, curr_row: &mut usize) {
-        let (next_row, circuit_gates) = Self::create_range_check(*curr_row);
-        *curr_row = next_row;
-        gates.extend_from_slice(&circuit_gates);
     }
 
     /// Verify the witness against a range check (related) circuit gate