--- a/kimchi/src/circuits/polynomials/range_check/witness.rs
+++ b/kimchi/src/circuits/polynomials/range_check/witness.rs
@@ -1,17 +1,22 @@
 //! Range check witness computation
 
 use ark_ff::PrimeField;
-use o1_utils::ForeignElement;
+use num_bigint::BigUint;
+use num_integer::Integer;
+use o1_utils::field_helpers::BigUintFieldHelpers;
+use o1_utils::{FieldHelpers, ForeignElement};
 use std::array;
 
 use crate::circuits::witness::Variables;
+use crate::variable_map;
 use crate::{
     circuits::{
         polynomial::COLUMNS,
-        witness::{init_row, ConstantCell, CopyBitsCell, CopyCell, VariableCell, WitnessCell},
+        witness::{init_row, CopyBitsCell, CopyCell, VariableCell, WitnessCell},
     },
     variables,
 };
+use o1_utils::foreign_field::BigUintForeignFieldHelpers;
 
 /// Witness layout
 ///   * The values and cell contents are in little-endian order.
@@ -33,16 +38,17 @@ fn layout<F: PrimeField>() -> Vec<[Box<dyn WitnessCell<F>>; COLUMNS]> {
         /* row 3, RangeCheck1 row */
         [
             VariableCell::create("v2"),
+            VariableCell::create("v12"), // optional
             /* 2-bit crumbs (placed here to keep lookup pattern */
             /*               the same as RangeCheck0) */
             CopyBitsCell::create(2, 0, 86, 88),
-            CopyBitsCell::create(2, 0, 84, 86),
             /* 12-bit plookups */
-            CopyBitsCell::create(2, 0, 72, 84),
-            CopyBitsCell::create(2, 0, 60, 72),
-            CopyBitsCell::create(2, 0, 48, 60),
-            CopyBitsCell::create(2, 0, 36, 48),
+            CopyBitsCell::create(2, 0, 74, 86),
+            CopyBitsCell::create(2, 0, 62, 74),
+            CopyBitsCell::create(2, 0, 50, 62),
+            CopyBitsCell::create(2, 0, 38, 50),
             /* 2-bit crumbs */
+            CopyBitsCell::create(2, 0, 36, 38),
             CopyBitsCell::create(2, 0, 34, 36),
             CopyBitsCell::create(2, 0, 32, 34),
             CopyBitsCell::create(2, 0, 30, 32),
@@ -50,11 +56,10 @@ fn layout<F: PrimeField>() -> Vec<[Box<dyn WitnessCell<F>>; COLUMNS]> {
             CopyBitsCell::create(2, 0, 26, 28),
             CopyBitsCell::create(2, 0, 24, 26),
             CopyBitsCell::create(2, 0, 22, 24),
-            CopyBitsCell::create(2, 0, 20, 22),
         ],
         /* row 4, Zero row */
         [
-            ConstantCell::create(F::zero()),
+            CopyBitsCell::create(2, 0, 20, 22),
             /* 2-bit crumbs (placed here to keep lookup pattern */
             /*               the same as RangeCheck0) */
             CopyBitsCell::create(2, 0, 18, 20),
@@ -115,7 +120,38 @@ pub fn create_multi<F: PrimeField>(v0: F, v1: F, v2: F) -> [Vec<F>; COLUMNS] {
 
     init_row(&mut witness, 0, 0, &layout, &variables!(v0));
     init_row(&mut witness, 0, 1, &layout, &variables!(v1));
-    init_row(&mut witness, 0, 2, &layout, &variables!(v2));
+    init_row(
+        &mut witness,
+        0,
+        2,
+        &layout,
+        &variable_map!("v2" => v2, "v12" => F::zero()),
+    );
+    init_row(&mut witness, 0, 3, &layout, &variables!());
+
+    witness
+}
+
+/// Create a multi range check witness from two limbs: v01 (176 bits), v2 (88 bits),
+/// where v2 is the most significant limb and v01 is the least significant limb
+pub fn create_multi_compact<F: PrimeField>(v01: F, v2: F) -> [Vec<F>; COLUMNS] {
+    let layout = layout();
+    let mut witness: [Vec<F>; COLUMNS] = array::from_fn(|_| vec![F::zero(); 4]);
+
+    let (v1, v0) = v01.to_biguint().div_rem(&BigUint::two_to_limb());
+    let v0: F = v0.to_field().expect("failed to convert to field element");
+    let v1: F = v1.to_field().expect("failed to convert to field element");
+
+    init_row(&mut witness, 0, 0, &layout, &variable_map!("v0" => v2));
+    init_row(&mut witness, 0, 1, &layout, &variable_map!("v1" => v0));
+
+    init_row(
+        &mut witness,
+        0,
+        2,
+        &layout,
+        &variable_map!("v2" => v1, "v12" => v01),
+    );
     init_row(&mut witness, 0, 3, &layout, &variables!());
 
     witness
@@ -124,6 +160,11 @@ pub fn create_multi<F: PrimeField>(v0: F, v1: F, v2: F) -> [Vec<F>; COLUMNS] {
 /// Create a multi range check witness from limbs
 pub fn create_multi_limbs<F: PrimeField>(limbs: &[F; 3]) -> [Vec<F>; COLUMNS] {
     create_multi(limbs[0], limbs[1], limbs[2])
+}
+
+/// Create a multi range check witness from compact limbs
+pub fn create_multi_compact_limbs<F: PrimeField>(limbs: &[F; 2]) -> [Vec<F>; COLUMNS] {
+    create_multi_compact(limbs[0], limbs[1])
 }
 
 /// Create a single range check witness
@@ -137,7 +178,7 @@ pub fn create<F: PrimeField>(v0: F) -> [Vec<F>; COLUMNS] {
     witness
 }
 
-/// Extend an existing witness with a multi-range-check gadget for foreign field element
+/// Extend an existing witness with a multi-range-check gadget for three 88-bit values: v0, v1 and v2
 pub fn extend_multi<F: PrimeField>(witness: &mut [Vec<F>; COLUMNS], v0: F, v1: F, v2: F) {
     let limbs_witness = create_multi(v0, v1, v2);
     for col in 0..COLUMNS {
@@ -145,9 +186,26 @@ pub fn extend_multi<F: PrimeField>(witness: &mut [Vec<F>; COLUMNS], v0: F, v1: F
     }
 }
 
-/// Extend an existing witness with a multi-range-check gadget for foreign field element
+/// Extend and existing witness with a multi range check witness for two limbs: v01 (176 bits), v2 (88 bits),
+/// where v2 is the most significant limb and v01 is the least significant limb
+pub fn extend_multi_compact<F: PrimeField>(witness: &mut [Vec<F>; COLUMNS], v01: F, v2: F) {
+    let limbs_witness = create_multi_compact(v01, v2);
+    for col in 0..COLUMNS {
+        witness[col].extend(limbs_witness[col].iter())
+    }
+}
+
+/// Extend an existing witness with a multi-range-check gadget for limbs
 pub fn extend_multi_limbs<F: PrimeField>(witness: &mut [Vec<F>; COLUMNS], limbs: &[F; 3]) {
     let limbs_witness = create_multi_limbs(limbs);
+    for col in 0..COLUMNS {
+        witness[col].extend(limbs_witness[col].iter())
+    }
+}
+
+/// Extend an existing witness with a multi-range-check gadget for compact limbs
+pub fn extend_multi_compact_limbs<F: PrimeField>(witness: &mut [Vec<F>; COLUMNS], limbs: &[F; 2]) {
+    let limbs_witness = create_multi_compact_limbs(limbs);
     for col in 0..COLUMNS {
         witness[col].extend(limbs_witness[col].iter())
     }