This PR adds the ability to constrain 2-limbs format (i.e. compact limbs) values with the multi-range-check gadget.

Normally the gadget can only constrain 3 limbs (`v0, v1, v2`) that are copied to column 0 of the first 3 gates of the gadget, like this

| 0  `RangeCheck0` | 1   `RangeCheck0`  | 2    `RangeCheck1` | 4 `Zero`  |
| -- | -- | -- | -- | 
| `v0` (copy) | `v1` (copy) | `v2` (copy) | ... |
| ... | ... | ... | ... |

This change allows one to copy the limbs in compact form: `v01`, `v2` to the gadget in order like this

| 0   `RangeCheck0`  | 1   `RangeCheck0`  |  2    `RangeCheck1` | 4 `Zero`  |
| -- | -- | -- | -- | 
| `v2` (copy) | `v0` | `v1` | ... |
| ... | ... | `v01` (copy) | ... |
| ... | ... | ... | ... |

where $v_{01} = v_0 + 2^{88} \cdot v_1$.

The change is comprised of two changes

* Move a couple witness elements around in `RangeCheck1`
* Add support for compact limbs to `RangeCheck0` gate (new optional mode of operation based on `coeff[0]`)

Closes #900

**Performance** 

```
Before
- time to create prover index: 56s
- time to create proof: 45s
- time to verify: 763ms
```

```
After
- time to create prover index: 56s
- time to create proof: 46s
- time to verify: 787ms
```